maany refactoring and stuff

+++ src/main/java/com/dominiknagy/tinyexpense/TinyExpense/implementations/CategoryServiceImpl.java
@@ -6,7 +6,10 @@ import com.dominiknagy.tinyexpense.TinyExpense.requests.CreateCategoryRequest;
 import com.dominiknagy.tinyexpense.TinyExpense.responses.CategoryResponse;
 import com.dominiknagy.tinyexpense.TinyExpense.services.CategoryService;
 import com.dominiknagy.tinyexpense.TinyExpense.services.UserService;
+import com.dominiknagy.tinyexpense.TinyExpense.utility.Mapper;
+import com.dominiknagy.tinyexpense.TinyExpense.utility.UserUtils;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.access.AuthorizationServiceException;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
@@ -17,19 +20,25 @@ import java.util.List;
 public class CategoryServiceImpl implements CategoryService {
 
     private final CategoryRepository categoryRepository;
-    private final UserService userService;
 
     @Override
-    public Category retrieveCategory(long expenseCategoryId) {
-        return categoryRepository.findExpenseCategoryById(expenseCategoryId).orElse(null);
+    public Category retrieveCategory(long categoryId) {
+        return categoryRepository.findCategoryById(categoryId).orElseThrow();
     }
 
     @Override
-    public List<CategoryResponse> retrieveCategories(String userEmail) {
+    public CategoryResponse retrieveCategoryAsResponse(long categoryId) {
+        Category category = categoryRepository.findCategoryById(categoryId).orElseThrow();
+        if (category.getUser() == UserUtils.authedUser())
+           return Mapper.mapCategoryResponse(categoryRepository.findCategoryById(categoryId).orElseThrow());
+        else throw new AuthorizationServiceException("User not authorized for this resource");
+    }
+
+    @Override
+    public List<CategoryResponse> retrieveCategories() {
         List<CategoryResponse> categoryResponses = new ArrayList<>();
 
-        for (Category category : categoryRepository.findCategoriesByUser(userService.retrieveUserByEmail(userEmail))) {
-
+        for (Category category : categoryRepository.findCategoriesByUser(UserUtils.authedUser())) {
             CategoryResponse categoryResponse = new CategoryResponse();
             categoryResponse.setId(category.getId());
             categoryResponse.setName(category.getCategoryName());
@@ -41,19 +50,20 @@ public class CategoryServiceImpl implements CategoryService {
     }
 
     @Override
-    public Category createCategory(
-            CreateCategoryRequest createCategoryRequest, String userEmail) {
-
+    public CategoryResponse createCategory(CreateCategoryRequest createCategoryRequest) {
         Category category = new Category();
         category.setCategoryName(createCategoryRequest.getCategoryName());
-        category.setUser(userService.retrieveUserByEmail(userEmail));
+        category.setUser(UserUtils.authedUser());
         category.setColor(createCategoryRequest.getColor());
 
-        return categoryRepository.save(category);
+        return Mapper.mapCategoryResponse(categoryRepository.save(category));
     }
 
     @Override
     public void deleteCategory(long categoryId) {
-        categoryRepository.deleteById(categoryId);
+        Category category = categoryRepository.findCategoryById(categoryId).orElseThrow();
+        if (category.getUser() == UserUtils.authedUser())
+            categoryRepository.deleteById(categoryId);
+        else throw new AuthorizationServiceException("User not authorized");
     }
 }
